# src/app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API MemoBelc",
    "description": "Documentação da API MemoBelc. Use **Authorize** para informar o token Bearer e testar rotas protegidas."
  },
  "host": "127.0.0.1:5000",
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Token JWT: use o valor 'Bearer <seu_token>'"
    }
  },
  "security": [],
...
  "paths": {
...
    "/billing/checkout": {
      "post": {
        "tags": ["Billing"],
        "summary": "Iniciar checkout Asaas (web) ou Google Play (Android)",
        "security": [{ "Bearer": [] }],
        "parameters": [{
          "name": "body",
          "in": "body",
          "schema": {
            "type": "object",
            "required": ["product_type", "product_id"],
            "properties": {
              "product_type": { "type": "string", "enum": ["plan", "book", "bundle"] },
              "product_id": { "type": "string" },
              "platform": { "type": "string", "enum": ["web", "android", "ios"] },
              "billing_type": { "type": "string" },
              "coupon_code": { "type": "string" }
            }
          }
        }],
        "responses": { "200": { "description": "URL de checkout ou SKU" } }
      }
    },
    "/billing/me": {
      "get": {
        "tags": ["Billing"],
        "summary": "Assinatura e entitlements do usuário",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Estado da assinatura" } }
      }
    },
    "/billing/payments": {
      "get": {
        "tags": ["Billing"],
        "summary": "Histórico de pagamentos do usuário",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Pagamentos" } }
      }
    },
    "/billing/cancel": {
      "post": {
        "tags": ["Billing"],
        "summary": "Cancelar assinatura atual",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Cancelada" } }
      }
    },
    "/billing/change-plan": {
      "post": {
        "tags": ["Billing"],
        "summary": "Upgrade ou downgrade de plano (mesmo provedor)",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Plano atualizado" } }
      }
    },
...
    "/billing/google/verify": {
      "post": {
        "tags": ["Billing"],
        "summary": "Validar compra Google Play",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Compra validada" } }
      }
    },
...
    "/admin/billing/subscriptions": {
      "get": {
        "tags": ["AdminBilling"],
        "summary": "Listar assinaturas com filtros",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Assinaturas" } }
      }
    },
    "/admin/billing/payments": {
      "get": {
        "tags": ["AdminBilling"],
        "summary": "Listar compras e transações",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Pagamentos" } }
      }
    },
    "/admin/billing/grants": {
      "post": {
        "tags": ["AdminBilling"],
        "summary": "Liberar plano, livro, conjunto ou serviço manualmente",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Concedido" } }
      }
    },
    "/admin/billing/external-sales": {
      "get": {
        "tags": ["AdminBilling"],
        "summary": "Listar vendas externas",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Vendas" } }
      },
      "post": {
        "tags": ["AdminBilling"],
        "summary": "Registrar venda externa e conceder acesso",
        "security": [{ "Bearer": [] }],
        "responses": { "201": { "description": "Registrada" } }
      }
    },
    "/admin/billing/access": {
      "get": {
        "tags": ["AdminBilling"],
        "summary": "Regras de visibilidade dos serviços",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Regras" } }
      }
    }
  }
}